4.15 - 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", {
@@ -11,14 +12,32 @@ blogsRouter.get("/", async (request, response) => {
   response.json(blogs);
 });
 
+blogsRouter.get("/:id", async (request, response) => {
+  const blog = await Blog.findById(request.params.id).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
+
+  if (blog) {
+    response.json(blog);
+  } else {
+    response.status(404).json({ error: "blog not found" }).end();
+  }
+});
+
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
 
   if (!body.title || !body.url) {
     return response.status(400).end();
   }
-
-  const user = await User.findById(body.userId);
 
   const blog = new Blog({
     title: body.title,
@@ -35,8 +54,22 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (user.id.toString() === blogToDelete.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    response
+      .status(403)
+      .json({ error: "only the creator can delete this blog" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
